refactor: build fix and renamed files
