+++ frontend/src/components/Logs/CategoryHeading/CategoryHeading.styles.scss
@@ -1,4 +1,4 @@
-@use 'scss/main';
+@use '/src/scss/main.scss';
 
 .category-heading-text {
     font-size: main.$size-font-regular;
