+++ frontend/src/components/Logs/CategoryHeading/index.tsx
@@ -1,0 +1,3 @@
+import CategoryHeading from './CategoryHeading';
+
+export default CategoryHeading;
