+++ frontend/webpack.config.js
@@ -143,15 +143,7 @@ const config = {
 					// Translates CSS into CommonJS
 					cssLoader,
 					// Compiles Sass to CSS
-					{
-						loader: sassLoader,
-						options: {
-							sassOptions: {
-								indentWidth: 4,
-								includePaths: ['./src'],
-							},
-						},
-					},
+					sassLoader,
 				],
 			},
 		],
